requirements.txt
Django>=5.2.5
psycopg2-binary>=2.9.9
django-environ>=0.11.2
python-decouple>=3.8
djangorestframework>=3.14.0
django-filter>=23.5
faker>=22.0.0
coreapi>=2.0.0
djangorestframework-simplejwt>=5.3.0
drf-yasg>=1.21.7
